Create Lobby Page
Implemented the automatic scroll feature for the lobby page. Currently there are two ```playerCards``` which is part of the HTML to display the profile pictures and names for each player and should be removed from the HTML file. Some lines of the CSS for each player card should also be removed since it will need to be alternating left and right.