Add menu to all pages
Added a settings menu that could be opened by clicking the logo at the top right hand corner. The menu gives the user the options to join a new game and to go back to the home screen. Styling for the menu is located in ```shared.css``` and the functionality is in ```menu.js```. Styling was also adjusted to ensure that it displays correctly on mobile.